keyword()

+++ src/main/java/com/newjeans/quickboard/service/KeywordService.java
@@ -1,12 +1,16 @@
 package com.newjeans.quickboard.service;
 
 import com.newjeans.quickboard.config.BaseException;
-import com.newjeans.quickboard.config.BaseResponseStatus;
+import com.newjeans.quickboard.domain.subscribe.Subscribe;
+import com.newjeans.quickboard.domain.User.User;
 import com.newjeans.quickboard.domain.keyword.Keyword;
 import com.newjeans.quickboard.domain.keyword.KeywordRepository;
+import com.newjeans.quickboard.domain.User.UserRepository;
+import com.newjeans.quickboard.domain.subscribe.SubscribeRepository;
 import com.newjeans.quickboard.web.dto.KeywordSaveRequestDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import static com.newjeans.quickboard.config.BaseResponseStatus.DATABASE_ERROR;
 
@@ -16,36 +20,53 @@ public class KeywordService {
 
     private final KeywordRepository keywordRepository;
 
-    public Long save(KeywordSaveRequestDto requestDto) {
+    private final UserRepository userRepository;
 
-        if(checkKeywordExist(requestDto.getKeyword())) {
-            requestDto.KeywordplusRequestDto(requestDto.getSubscriberCount());
-            return keywordRepository.findByKeyword(requestDto.getKeyword()).getId();
-
-        } else
-            return keywordRepository.save(requestDto.toEntity()).getId();
+    private final SubscribeRepository subscribeRepository;
+    @Transactional
+    public void save(String uuid, KeywordSaveRequestDto requestDto) throws BaseException {
+        try {
+            if(checkKeywordExist(requestDto.getKeyword())) {
+                Keyword keyword = keywordRepository.findByKeyword(requestDto.getKeyword());
+                keyword.plusSubscribers();
+                User user = userRepository.getReferenceByUuid(uuid);
+                Keyword findKeyword = keywordRepository.findByKeyword(requestDto.getKeyword());
+                Subscribe subscribe = Subscribe.builder().user(user).keyword(findKeyword).build();
+                subscribeRepository.save(subscribe);
+            } else {
+                User user = userRepository.getReferenceByUuid(uuid);
+                Keyword saveKeyword = Keyword.builder().keyword(requestDto.getKeyword()).build();
+                keywordRepository.save(saveKeyword);
+                Keyword findKeyword = keywordRepository.findByKeyword(requestDto.getKeyword());
+                findKeyword.plusSubscribers();
+                Subscribe subscribe = Subscribe.builder().user(user).keyword(findKeyword).build();
+                subscribeRepository.save(subscribe);
+            }
+        }catch (Exception exception){
+            exception.printStackTrace();
+            throw new BaseException(DATABASE_ERROR);
+        }
     }
 
-
-    public void deleteAllByKeyword(String keyword) throws BaseException {
+    public void deleteAllByKeyword(String uuid, String keyword) throws BaseException {
         try {
+            User user = userRepository.getReferenceByUuid(uuid);
             Keyword deleteKeyword = keywordRepository.getReferenceByKeyword(keyword);
-            keywordRepository.delete(deleteKeyword);
+
+            subscribeRepository.deleteAllByUser(user);
+            System.out.println("@@@@@@@@@@@@@@@@@");
+            deleteKeyword.minusSubscribers();
+            if(deleteKeyword.getSubscribersCount()==0) {
+                keywordRepository.delete(deleteKeyword);
+            }
         } catch(Exception exception) {
             throw new BaseException(DATABASE_ERROR);
         }
     }
 
     public boolean checkKeywordExist(String keyword) {
-        return keywordRepository.existsKByKeyword(keyword);
+        return keywordRepository.existsByKeyword(keyword);
     }
-
-    public void plussubscribersCount(Keyword keyword) {
-        keyword.builder().subscribersCount(keyword.getSubscribersCount() + 1).build();
-    }
-
-
-
 }
